refactor: extract shared test helpers into common/mod.rs
Deduplicate has_github_token(), has_project_number(), test_config(), and
test_server_state() which were copied verbatim between integration.rs
and e2e_workflow.rs. Both test files now import from a shared
tests/common/mod.rs module to prevent drift.
